requirements.txt
Django==4.2.11
djangorestframework==3.15.1
django-filter==23.5
Pillow>=11.0.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
django-redis==5.4.0
